#944 rename app bar component
