+++ wls-gui-wahllokalsystem/src/components/wlsComponents/TheWlsAppBar.vue
